thêm, sửa xóa

+++ app/src/main/java/com/example/ecommerce/model/Review.kt
@@ -1,14 +1,14 @@
 package com.example.ecommerce.model
 
-import java.sql.Timestamp
+import com.google.firebase.Timestamp
 
 data class Review(
     val id: String,
     val productId: String,
     val userId: String,
-    val userName: String? = null, // Thêm để hiển thị tên người đánh giá
-    val userAvatarUrl: String? = null, // Thêm để hiển thị ảnh đại diện
-    val rating: Float,
+    val userName: String? = null,
+    val userAvatarUrl: String? = null,
+    val rating: Int,
     val comment: String?,
-    val createdAt: Timestamp // Đổi thành Timestamp
+    val createdAt: Timestamp
 )
